requirements.txt
python-dotenv
playwright
requests
openai
fastapi
uvicorn[standard]
pydantic-settings
jinja2
aiofiles
python-socks
apscheduler
httpx[socks]
Pillow
pyzbar
qrcode
pytest
pytest-asyncio
coverage
bcrypt
